Former-commit-id: ce7ef5b27158bc627e298a3e1f285e38238c86b9

+++ client/src/Components/Community/Board/index.tsx
@@ -22,7 +22,7 @@ export default function Board(): JSX.Element {
   const [isAuthor, setIsAuthor] = useState<Boolean>(false);
   const modifyProps = [data, item];
   const [cookies, setCookie] = useCookies(['plant-blog']);
-  console.log(data && data.content);
+
   // 게시글 번호에 맞는 글 불러오기
   useEffect(() => {
     async function getPost() {
@@ -50,23 +50,33 @@ export default function Board(): JSX.Element {
 
   // 게시글 삭제
   async function onDeleteHandler() {
-    try {
-      await CommunityApi.Community_Post(`blog/delete/${item}`, '').then(
-        response => {
-          if (response.data.new_token !== null) {
-            console.log('새로운 토큰이 도착했습니다!');
-            localStorage.removeItem('access_token');
-            localStorage.setItem('access_token', response.data.new_token);
-            navigate('/community');
-          } else {
-            console.log('뉴토큰이 없습니다.');
-            navigate('/community');
-          }
-        },
-      );
-    } catch (e) {
-      toast.error('본인만 삭제할 수 있습니다.', {
-        autoClose: 2500,
+    let reconfirmMessage: string | null = prompt(
+      '"삭제"를 입력하시면 글이 삭제됩니다.',
+    );
+
+    if (reconfirmMessage === '삭제') {
+      try {
+        await CommunityApi.Community_Post(`blog/delete/${item}`, '').then(
+          response => {
+            if (response.data.new_token !== null) {
+              console.log('새로운 토큰이 도착했습니다!');
+              localStorage.removeItem('access_token');
+              localStorage.setItem('access_token', response.data.new_token);
+              navigate('/community');
+            } else {
+              console.log('뉴토큰이 없습니다.');
+              navigate('/community');
+            }
+          },
+        );
+      } catch (e) {
+        toast.error('본인만 삭제할 수 있습니다.', {
+          autoClose: 2500,
+        });
+      }
+    } else {
+      toast.info('삭제가 취소되었습니다.', {
+        position: toast.POSITION.TOP_CENTER,
       });
     }
   }
